Experiment with a "blob mode".
In that mode you can expand your word by adding any letter that is adjacent to
any other letter in the blob.

Kinda neat, but it makes it impossible to remove letters from your accumulating
word, which is kinda pesky. Maybe I can figure out a way to do that.